+++ features/overlays/totemTracker.js
@@ -2,7 +2,7 @@ import settings from "../../settings";
 import { overlayCoordsData } from "../../data/overlayCoords";
 import { EntityArmorStand } from "../../constants/javaTypes";
 import { TIMER_SOUND_SOURCE, OFF_SOUND_MODE } from "../../constants/sounds";
-import { WHITE, GOLD, RED } from "../../constants/formatting";
+import { WHITE, RED, DARK_PURPLE } from "../../constants/formatting";
 import { isInSkyblock } from "../../utils/playerState";
 
 let remainingTotemTime; // Format examples: 01m 02s, 50s, 09s
@@ -12,6 +12,9 @@ const secondsBeforeExpiration = 10;
 
 register('step', () => trackTotemStatus()).setFps(1);
 register('renderOverlay', () => renderTotemOverlay());
+register("worldUnload", () => {
+    resetTotem();
+});
 
 function trackTotemStatus() {
     if ((!settings.alertOnTotemExpiresSoon && !settings.totemRemainingTimeOverlay) || !isInSkyblock()) {
@@ -24,8 +27,7 @@ function trackTotemStatus() {
         return name.includes('Owner:') && name.includes(currentPlayer);
     });
     if (playerTotemPosition && !hasPlayersTotem) {
-        playerTotemPosition = null;
-        remainingTotemTime = null;
+        resetTotem();
     }
 
 	entities.forEach(entity => {
@@ -41,7 +43,7 @@ function trackTotemStatus() {
                 remainingTotemTime = name.split('Remaining: ').pop();
 
                 if (settings.alertOnTotemExpiresSoon && remainingTotemTime && remainingTotemTime === `${secondsBeforeExpiration}s`) {
-                    Client.showTitle(`&cYour totem expires soon`, '', 1, 45, 1);
+                    Client.showTitle(`${DARK_PURPLE}Totem ${RED}expires soon`, '', 1, 30, 1);
 
                     if (settings.soundMode !== OFF_SOUND_MODE)
                     {
@@ -54,14 +56,19 @@ function trackTotemStatus() {
 }
 
 function renderTotemOverlay() {
-    if (!settings.totemRemainingTimeOverlay || !remainingTotemTime || !isInSkyblock()) {
+    if (!settings.totemRemainingTimeOverlay || !remainingTotemTime || remainingTotemTime === '00s' || !isInSkyblock()) {
         return;
     }
 
     const timerColor = !remainingTotemTime.includes('m') && remainingTotemTime.slice(0, -1) <= secondsBeforeExpiration ? RED : WHITE;
-    const overlayText = `${GOLD}Remaining totem time: ${timerColor}${remainingTotemTime}`;
+    const overlayText = `${DARK_PURPLE}Totem of Corruption: ${timerColor}${remainingTotemTime}`;
     const overlay = new Text(overlayText, overlayCoordsData.totemRemainingTimeOverlay.x, overlayCoordsData.totemRemainingTimeOverlay.y)
         .setShadow(true)
         .setScale(overlayCoordsData.totemRemainingTimeOverlay.scale);
     overlay.draw();
+}
+
+function resetTotem() {
+    playerTotemPosition = null;
+    remainingTotemTime = null;
 }